moving test helpers into separate file
